making progress on num.cr txlations

+++ tests/numpy/linspace.py
@@ -1,6 +1,7 @@
 import numpy as np
 
-x = np.linspace(2.0, 3.0, num=5)
+x = np.linspace(2.0, 3.0, num=5, dtype=np.float64)
+
 for a in x:
     print(a)
